test: use ktor test support for running a model-server in `TypedModelQLTest`
The previous solution with this `ci.sh` script starting a model server process is terrible for
debugging. Have to migrate all the other tests as well, but this one I just had to debug.
